Adding FAB and icon resource.
Adding a vector asset for the + sign in the FAB.
Added FAB in the XML view. Using Coordinator layout because it automatically gives behavior benefits when using it with FAB. The view will comply with Snackbars, AppBarLayout, and BottomSheetBehavior. Refer to https://material.io/develop/android/components/floating-action-button/ for more info.